admin: factor out honey badger service
